+++ src/interface/index.js
@@ -1,3 +1,4 @@
 
 export * from "./initializeElement";
 export * from "./jumpToSlice";
+export * from "./tools";