+++ src/interface/initializeElement.js
@@ -1,22 +1,8 @@
 
 import {assert} from "chai";
+
 import {renderElement} from "../rendering";
-import {getElementImageContext, ImageContext, setElementImageContext} from "../context";
-import {jumpToSlice} from "./jumpToSlice";
-
-function wheelEventHandler(event) {
-
-  const canvasElement = event.target;
-  const imageContext = getElementImageContext(canvasElement);
-
-  const sliceDelta = (event.deltaY > 0) ? 1 : -1;
-  const newSlice = imageContext.slice + sliceDelta;
-
-  if (newSlice >= 0 && newSlice < imageContext.dimensions[0]) {
-
-    jumpToSlice(canvasElement, newSlice);
-  }
-}
+import {ImageContext, setElementImageContext} from "../context";
 
 /**
  * @param {!HTMLCanvasElement} canvasElement
@@ -26,9 +12,9 @@ function wheelEventHandler(event) {
  */
 export function initializeElement(canvasElement, data, dimensions, slice) {
 
-  console.info(`Initializing ostrich on element:`, canvasElement);
-
   assert.typeOf(canvasElement, HTMLCanvasElement.name);
+
+  console.debug(`Initializing ostrich on element:`, canvasElement);
 
   const ctx = canvasElement.getContext('2d', {
 
@@ -46,8 +32,6 @@ export function initializeElement(canvasElement, data, dimensions, slice) {
   // todo: (can be removed as soon as transform is implemented for zoom/pan)
   canvasElement.width = imageContext.dimensions[1];
   canvasElement.height = imageContext.dimensions[2];
-
-  canvasElement.addEventListener('wheel', wheelEventHandler);
 
   canvasElement.dispatchEvent(new Event('ostrich.initialized', {
     bubbles: true,