+++ src/interface/tools.js
@@ -1,0 +1,37 @@
+
+import {assert} from "chai";
+
+import {getElementImageContext} from "../context";
+import {Tool} from "../tools";
+
+export function activateTool(canvasElement, tool) {
+
+  assert.instanceOf(tool, Tool);
+
+  const imageContext = getElementImageContext(canvasElement);
+  imageContext.addTool(tool);
+
+  tool.activate(canvasElement);
+
+  canvasElement.dispatchEvent(new CustomEvent('ostrich.toolActivated', {
+    detail: {
+      tool: tool,
+    },
+  }));
+}
+
+export function deactivateTool(canvasElement, name) {
+
+  const imageContext = getElementImageContext(canvasElement);
+  const tool = imageContext.getTool(name);
+
+  tool.deactivate(canvasElement);
+
+  imageContext.removeTool(name);
+
+  canvasElement.dispatchEvent(new CustomEvent('ostrich.toolDeactivated', {
+    detail: {
+      tool: tool,
+    },
+  }));
+}