+++ src/tools/Tool.js
@@ -1,0 +1,17 @@
+
+export class Tool {
+
+  /**
+   * @param {HTMLCanvasElement} canvasElement
+   */
+  activate(canvasElement) {
+
+  }
+
+  /**
+   * @param {HTMLCanvasElement} canvasElement
+   */
+  deactivate(canvasElement) {
+
+  }
+}