+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## [unreleased]
 
-- Add: Added events `ostrich.initialized`, `ostrich.rendered` and `ostrich.sliceChanged`
-- Add: Added `ostrich.jumpToSlice()`
+- Add: Added events `ostrich.initialized`, `ostrich.rendered`, `ostrich.sliceChanged`, `ostrich.activateTool`, `ostrich.deactivateTool`
+- Add: Added method `ostrich.jumpToSlice()`
+- Add: Added tool `ostrich.SliceScrollingTool`
 - Add: Implemented basic rendering
 
 ## 0.1.0